# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Skillset API",
    "version": "0.1.0",
    "description": "A self-hosted Registry for Agent Skills. This is the contract the CLI depends on; the web interface uses the same routes and has no private data path of its own.\n\n**Conventions.** Every JSON key on the wire is snake_case, which means wire keys line up with Postgres column names. TypeScript types, interfaces, and classes are camelCase, so conversion happens once at the boundary in the shared module rather than ad hoc per call site.\n\n**Artifact bytes never pass through this API.** Publishing returns a presigned upload target that the caller writes to directly, and reading an Artifact redirects to a short-lived presigned location. The API therefore validates the metadata a caller claims about a Skill, never the Artifact itself (ADR-0001)."
  },
  "servers": [{ "url": "/api" }],
  "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
...
  "paths": {
...
    "/skills": {
      "get": {
        "tags": ["skills"],
        "summary": "List, search, filter, and sort Skills",
        "description": "Reads from the `skill_directory` view (ticket 23, docs/data-model.md), not `skills` directly. Defaults to sorted by install count, descending. With `q`, results are additionally full-text matched over name, description, and publisher — `sort_by`/`sort_order` still govern the order even then; there is no separate relevance ranking. Stemming does not substring-match, and hyphenated names tokenise per word — both expected, not defects. Unauthenticated (ADR-0013).",
        "security": [],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "description": "Full-text search phrase, matched against name, description, and publisher. Supports quoted phrases and `-exclusions`. Omitted or empty returns the ordinary list.",
            "schema": { "type": "string" }
          },
          {
            "name": "tag_id",
            "in": "query",
            "required": false,
            "description": "Repeatable. Narrows to Skills carrying at least one of the given Tag ids (any-match). Omitted returns Skills regardless of Tags.",
            "schema": { "type": "array", "items": { "type": "string", "format": "uuid" } }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "description": "Governs ordering unconditionally, even with `q` set. Invalid values are rejected with a field-named validation error.",
            "schema": { "type": "string", "enum": ["installs", "updated_at"], "default": "installs" }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "description": "Invalid values are rejected with a field-named validation error.",
            "schema": { "type": "string", "enum": ["asc", "desc"], "default": "desc" }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "description": "Clamped to [1, 100] rather than rejected.",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 }
          },
          { "$ref": "#/components/parameters/Page" }
        ],
        "responses": {
          "200": {
            "description": "A page of Skills",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["items", "page", "page_size", "total"],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/SkillDirectoryEntry" }
                    },
                    "page": { "type": "integer" },
                    "page_size": { "type": "integer", "minimum": 1, "maximum": 100 },
                    "total": { "type": "integer" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationFailed" }
        }
      }
    },
    "/skills/{name}": {
      "parameters": [{ "$ref": "#/components/parameters/SkillName" }],
      "put": {
        "tags": ["skills"],
        "summary": "Publish a Skill",
        "description": "Create or replace, idempotent by name: publishing an existing Skill replaces it whoever published it first, and `published_by` and `published_at` become whoever published it last (ADR-0002).\n\nThe row is written first and an upload target returned. The caller then writes the Artifact to that target directly. Until it does, the Skill lists and reads but its Artifact cannot be retrieved.\n\n`name` comes from the path; the caller parses it out of `SKILL.md` frontmatter. The API validates it against the same rules the shared module applies, but does not read the Artifact to confirm they agree.\n\n`license`, `compatibility`, `metadata`, and `allowed_tools` are optional, but are validated exactly as strictly as `description` and `body` when present — any one of them failing rejects the whole publish, naming the specific field and rule (ADR-0009). Publishing always writes all four (a validated value, or null when absent), so a republish that no longer sets a field clears it. `tags` cannot be set through this endpoint at all — see `PUT /skills/{id}/tags`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["description", "body"],
                "properties": {
                  "description": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1024,
                    "description": "From the `SKILL.md` frontmatter."
                  },
                  "body": {
                    "type": "string",
                    "minLength": 1,
                    "description": "The `SKILL.md` content below the frontmatter. Rendered through an allowlist sanitiser, never trusted as markup."
                  },
                  "license": { "type": "string", "description": "From the SKILL.md frontmatter. Optional." },
                  "compatibility": {
                    "type": "string",
                    "maxLength": 500,
                    "description": "From the SKILL.md frontmatter. Optional."
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": { "type": "string" },
                    "description": "From the SKILL.md frontmatter, a mapping of string keys to string values. Optional."
                  },
                  "allowed_tools": {
                    "type": "string",
                    "description": "From the SKILL.md frontmatter. Optional; internal grammar unvalidated (experimental in the specification)."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The Skill row was created or replaced. Write the Artifact to `upload` next.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["skill", "upload"],
                  "properties": {
                    "skill": { "$ref": "#/components/schemas/Skill" },
                    "upload": { "$ref": "#/components/schemas/UploadTarget" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": {
            "description": "Readers cannot publish",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/skills/by-name/{name}": {
      "parameters": [{ "$ref": "#/components/parameters/SkillName" }],
      "get": {
        "tags": ["skills"],
        "summary": "Read a Skill by name",
        "description": "An exact-name read, for resolving `name` to `id` — the web keeps `/skills/<name>` as its own browser URL and needs this to land a reader there with nothing already cached (a bookmark, a shared link, a refresh). Not full-text search: `GET /skills?q=` ranks by recency and caps at one page, so it cannot guarantee finding an existing Skill by exact name, and Postgres's stopword handling can make it return nothing at all for a name that happens to be one. Unauthenticated (ADR-0013).",
        "security": [],
        "responses": {
          "200": {
            "description": "The Skill",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Skill" } } }
          },
          "404": { "$ref": "#/components/responses/SkillNotFound" }
        }
      }
    },
...
    "/skills/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/SkillId" }],
      "get": {
        "tags": ["skills"],
        "summary": "Read a Skill",
        "description": "An exact-id read. Does not go through full-text search. The web keeps `/skills/<name>` as its own browser URL and resolves `name` to `id` itself before calling this — see `GET /skills/by-name/{name}`. Unauthenticated (ADR-0013).",
        "security": [],
        "responses": {
          "200": {
            "description": "The Skill",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Skill" } } }
          },
          "404": { "$ref": "#/components/responses/SkillNotFound" }
        }
      },
      "delete": {
        "tags": ["skills"],
        "summary": "Delete a Skill",
        "description": "Requires at least the admin role. Removes the row and the Artifact permanently — there is no version history and no soft delete. Recovery is bucket versioning in object storage (ADR-0002). Succeeds even when the Artifact was never uploaded.",
        "responses": {
          "204": { "description": "Deleted. The name is immediately available to publish again." },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": {
            "description": "Readers and writers cannot delete a Skill",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "404": { "$ref": "#/components/responses/SkillNotFound" }
        }
      }
    },
    "/skills/{id}/artifact": {
      "parameters": [{ "$ref": "#/components/parameters/SkillId" }],
      "get": {
        "tags": ["skills"],
        "summary": "Retrieve a Skill's Artifact",
        "description": "Records one Install, then redirects to a presigned read location valid for about a minute. Callers must follow redirects. Unauthenticated — anyone who can reach the Registry may retrieve an Artifact (ADR-0013). The response body is the zip, named after the Skill. A caller that sends `Accept: application/json` gets the Skill itself back instead alongside `url` to fetch the Artifact from, rather than a redirect; this is for a caller (the web app's own Download control) that needs to know the request succeeded before navigating there itself. Either representation records the same one Install. `installs` on that Skill reflects the last periodic refresh, not necessarily this request's own Install (ADR-0012).",
        "security": [],
        "parameters": [
          {
            "name": "Accept",
            "in": "header",
            "required": false,
            "schema": { "type": "string", "enum": ["application/json"] },
            "description": "Request the Skill-plus-url representation instead of the default redirect."
          }
        ],
        "responses": {
          "200": {
            "description": "The Skill, plus a presigned url to fetch the Artifact from — returned instead of redirecting when the request sent `Accept: application/json`",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Skill" },
                    {
                      "type": "object",
                      "required": ["url"],
                      "properties": {
                        "url": {
                          "type": "string",
                          "format": "uri",
                          "description": "Short-lived presigned read URL for the Artifact"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "302": {
            "description": "Redirect to the presigned location — the default representation",
            "headers": {
              "Location": {
                "description": "Short-lived presigned read URL for the Artifact",
                "schema": { "type": "string", "format": "uri" }
              }
            }
          },
          "404": {
            "description": "No such Skill, or its Artifact was never uploaded",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/skills/{id}/tags": {
      "parameters": [{ "$ref": "#/components/parameters/SkillId" }],
      "put": {
        "tags": ["skills"],
        "summary": "Replace a Skill's Tags",
        "description": "Requires at least the writer role — the same bar as publishing. A full replace, not incremental: the given names become exactly this Skill's Tags, and any Tag it previously carried that isn't in the list is detached (never deleted from the catalog itself, ADR-0011). Each name is normalised (trimmed, lowercased) and resolved to an existing catalog Tag or a newly created one — this is the only way a brand new Tag comes into existence. A name repeated in the list, even under different casing, still resolves to one Tag.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tags"],
                "properties": {
                  "tags": {
                    "type": "array",
                    "items": { "type": "string" },
                    "description": "The full desired set of tag names, by name — not by id."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The Skill's resolved Tags after the replace",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["tags"],
                  "properties": {
                    "tags": { "type": "array", "items": { "$ref": "#/components/schemas/Tag" } }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": {
            "description": "Readers cannot edit a Skill's Tags",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "404": { "$ref": "#/components/responses/SkillNotFound" }
        }
      }
    },
...
    "/tags/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/TagId" }],
      "patch": {
        "tags": ["tags"],
        "summary": "Rename a Tag",
        "description": "Requires at least the admin role — stricter than replacing a Skill's Tags, because a rename reaches every Skill carrying this Tag, not just one (ADR-0011). Every reference is by id, so nothing but this Tag's `name` column changes. Conflict detection relies on the database's own unique constraint, so two concurrent renames to the same name can't both succeed.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 32,
                    "pattern": "^[a-z0-9]([a-z0-9-]*[a-z0-9])?$"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The renamed Tag",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Tag" } } }
          },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": {
            "description": "Only Admins may rename a Tag",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "404": { "$ref": "#/components/responses/TagNotFound" },
          "409": {
            "description": "A different Tag already holds this name",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Signed session cookie, http-only, secure, strictly same-site. Carries only the User's id; the role is resolved from the database on every request (ADR-0005). Cannot be revoked before expiry."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A Token, used by the CLI. Looked up by hash on every request; revocable immediately."
      }
    },
...
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": { "type": "integer", "minimum": 1, "default": 1 }
      },
      "SkillName": {
        "name": "name",
        "in": "path",
        "required": true,
        "description": "The Skill's name, from its `SKILL.md` frontmatter.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 64,
          "pattern": "^[a-z0-9]([a-z0-9-]*[a-z0-9])?$"
        }
      },
      "SkillId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "The Skill's id.",
        "schema": { "type": "string", "format": "uuid" }
      },
      "TagId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "The Tag's id.",
        "schema": { "type": "string", "format": "uuid" }
      },
...
    },
    "responses": {
...
      "Unauthenticated": {
        "description": "No valid session or Token. The web interface treats this as an expired session, discards cached state, and returns to login.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
      "SkillNotFound": {
        "description": "No Skill by that id",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "TagNotFound": {
        "description": "No Tag by that id",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
      "ValidationFailed": {
        "description": "The request failed a validation rule. `code` names the rule and `field` names what broke it, so the CLI and the web interface can report the same specific reason.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    },
    "schemas": {
...
      "Publisher": {
        "type": "object",
        "description": "Who last published a Skill. `email` is a snapshot taken at publish time and is always present; `user_id` and the names are null once that User has been removed.",
        "required": ["user_id", "email", "first_name", "last_name"],
        "properties": {
          "user_id": { "type": ["string", "null"], "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "first_name": { "type": ["string", "null"] },
          "last_name": { "type": ["string", "null"] }
        }
      },
      "SkillDirectoryEntry": {
        "type": "object",
        "description": "One row of `GET /skills` (ticket 23), sourced from the `skill_directory` view — deliberately not SkillSummary: `published_by_name` in place of the full Publisher object, and `updated_at` in place of `published_at`. `GET /skills/{id}` and `by-name/{name}` are unaffected and still return SkillSummary/Skill.",
        "required": ["id", "name", "description", "published_by_name", "updated_at", "installs", "tags"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": {
            "type": "string",
            "maxLength": 64,
            "pattern": "^[a-z0-9]([a-z0-9-]*[a-z0-9])?$"
          },
          "description": { "type": "string", "maxLength": 1024 },
          "published_by_name": {
            "type": "string",
            "description": "A display-name snapshot taken at publish time (docs/data-model.md) — does not update if that User later renames themselves."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "When this Skill was last published, republished, or otherwise changed — not when it was first published (`published_at`, only on the detail read)."
          },
          "installs": {
            "type": "integer",
            "minimum": 0,
            "description": "Same field, same lag caveat as SkillSummary.installs (ADR-0012)."
          },
          "tags": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Tag" }
          }
        }
      },
...
      "SkillSummary": {
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "license",
          "compatibility",
          "metadata",
          "allowed_tools",
          "tags",
          "published_by",
          "published_at",
          "installs"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Stable identity for reading, deleting, and downloading this Skill — generated once, on first publish, and never changes across republishes of the same name. `name` remains how a Skill is published."
          },
          "name": {
            "type": "string",
            "maxLength": 64,
            "pattern": "^[a-z0-9]([a-z0-9-]*[a-z0-9])?$"
          },
          "description": { "type": "string", "maxLength": 1024 },
          "license": {
            "type": ["string", "null"],
            "description": "From the SKILL.md frontmatter. Null when the Skill never set it."
          },
          "compatibility": {
            "type": ["string", "null"],
            "maxLength": 500,
            "description": "From the SKILL.md frontmatter. Null when the Skill never set it."
          },
          "metadata": {
            "type": ["object", "null"],
            "additionalProperties": { "type": "string" },
            "description": "From the SKILL.md frontmatter, a mapping of string keys to string values. Null when the Skill never set it."
          },
          "allowed_tools": {
            "type": ["string", "null"],
            "description": "From the SKILL.md frontmatter. Null when the Skill never set it."
          },
          "tags": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Tag" },
            "description": "Registry-owned, never parsed from or written into SKILL.md (ADR-0008), and never set by publishing — see `PUT /skills/{id}/tags`. Each Tag is drawn from the registry-wide catalog (ADR-0011); empty until something attaches one."
          },
          "published_by": { "$ref": "#/components/schemas/Publisher" },
          "published_at": { "type": "string", "format": "date-time" },
          "installs": {
            "type": "integer",
            "minimum": 0,
            "description": "How many times this Skill has been installed — a download today, and eventually a CLI install too. 0 until the first one is ever recorded, never null. Reflects the last periodic refresh of the underlying materialized view (ADR-0012), not necessarily every Install recorded so far — this can lag reality by a short interval."
          }
        }
      },
      "Skill": {
        "allOf": [
          { "$ref": "#/components/schemas/SkillSummary" },
          {
            "type": "object",
            "required": ["body"],
            "properties": {
              "body": {
                "type": "string",
                "description": "The `SKILL.md` content below the frontmatter, as supplied by the publisher. Sanitise before rendering."
              }
            }
          }
        ]
      },
      "UploadTarget": {
        "type": "object",
        "description": "Where to write the Artifact. A presigned location on object storage — the bytes do not pass through this API.",
        "required": ["url", "method", "headers", "expires_in_seconds"],
        "properties": {
          "url": { "type": "string", "format": "uri" },
          "method": { "type": "string", "const": "PUT" },
          "headers": {
            "type": "object",
            "additionalProperties": { "type": "string" },
            "description": "Headers that must be sent verbatim for the presigned signature to verify."
          },
          "expires_in_seconds": { "type": "integer" }
        }
      },
...
      "Tag": {
        "type": "object",
        "description": "A catalog Tag: a stable id plus its current name, shared by every Skill that carries it — renaming one (PATCH /tags/{id}) is visible everywhere it's attached, since every reference is by id, not name (ADR-0011).",
        "required": ["id", "name"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": {
            "type": "string",
            "maxLength": 32,
            "pattern": "^[a-z0-9]([a-z0-9-]*[a-z0-9])?$",
            "description": "Always stored already lowercased and trimmed."
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string",
                "description": "A stable machine-readable identifier for what went wrong — for validation failures, the specific rule that failed."
              },
              "message": { "type": "string" },
              "field": {
                "type": "string",
                "description": "The field or file that broke the rule, when there is one."
              }
            }
          }
        }
      }
    }
  }
}